python_code/requirements.txt
aiogram==2.25.1
emoji==2.2.0
googletrans==3.1.0a0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
pyTelegramBotAPI==4.12.0

